requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
dj-database-url==0.4.0
Django==1.9.2
django-extensions==1.6.1
django-widget-tweaks==1.4.1
djangorestframework==3.4.4
gunicorn==19.4.5
ipython==4.2.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.4
Pillow==3.3.0
psycopg2==2.6.1
ptyprocess==0.5.1
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
virtualenv==15.0.2
whitenoise==2.0.6
